util: Move Packable impls to toplevel util
