+++ src/util.rs
@@ -1,6 +1,6 @@
 //! Various utility functions, structs and traits
 
-use crate::error::{Error, ErrorType};
+use crate::error::{Error, ErrorExt, ErrorType};
 use std::io::{self, Read, Write};
 
 pub mod architecture;
@@ -68,4 +68,65 @@ pub trait ReprU16 {
     fn from_u16(num: u16) -> Option<Self>
     where
         Self: Sized;
+}
+
+impl Packable for u8 {
+    fn pack<W: Write>(&self, output: &mut W) -> Result<(), Error> {
+        output.write(&[*self]).ctx(|| format!("Writing {self}"))?;
+
+        Ok(())
+    }
+}
+
+impl Unpackable for u8 {
+    fn unpack<R: Read>(input: &mut R) -> Result<Option<Self>, Error> {
+        let mut buf = [0u8; 1];
+        let x = input.read(&mut buf).ctx(|| "Read u8".to_owned())?;
+        Ok(match x {
+            1 => Some(buf[0]),
+            _ => None,
+        })
+    }
+}
+
+impl Packable for u16 {
+    fn pack<W: Write>(&self, output: &mut W) -> Result<(), Error> {
+        output
+            .write(&self.to_le_bytes())
+            .ctx(|| format!("Writing {self}"))?;
+
+        Ok(())
+    }
+}
+
+impl Unpackable for u16 {
+    fn unpack<R: Read>(input: &mut R) -> Result<Option<Self>, Error> {
+        let mut buf = [0u8; 2];
+        let x = input.read(&mut buf).ctx(|| "Read u16".to_owned())?;
+        Ok(match x {
+            2 => Some(Self::from_le_bytes(buf)),
+            _ => None,
+        })
+    }
+}
+
+impl Packable for u32 {
+    fn pack<W: Write>(&self, output: &mut W) -> Result<(), Error> {
+        output
+            .write(&self.to_le_bytes())
+            .ctx(|| format!("Writing {self}"))?;
+
+        Ok(())
+    }
+}
+
+impl Unpackable for u32 {
+    fn unpack<R: Read>(input: &mut R) -> Result<Option<Self>, Error> {
+        let mut buf = [0u8; 4];
+        let x = input.read(&mut buf).e_context(|| "Read u32".to_owned())?;
+        Ok(match x {
+            4 => Some(Self::from_le_bytes(buf)),
+            _ => None,
+        })
+    }
 }
